Vuepress 2.x version
Updated Vuepress 2.x version.
Fixed #1239 #1238 #1211
