+++ .gitignore
@@ -1,3 +1,5 @@
 /docs/.vuepress/dist
+/docs/.vuepress/.cache
+/docs/.vuepress/.temp
 /node_modules
 /.vs
