+++ README.md
@@ -1,4 +1,10 @@
-# Wasabi Wallet 2.0 Documentation
+<p align="center">
+  <a href="https://docs.wasabiwallet.io">
+    <img src="https://user-images.githubusercontent.com/51679301/173640921-827d7f4f-5bb8-432b-9c53-7146a7ca6763.png" alt="Under Construction" title="Documentation Under Construction" class="Image" />
+  </a>
+</p>
+
+# Wasabi Wallet 2.0 Documentation is under construction!
 
 [Wasabi Wallet](https://wasabiwallet.io) is an open-source, non-custodial, privacy focused Bitcoin wallet for desktop. It implements a trustless privacy technique, called [WabiSabi CoinJoin](https://eprint.iacr.org/2021/206).
