+++ docs/.vuepress/styles/palette.styl
