+++ docs/.vuepress/variables.js
@@ -1,7 +1,7 @@
 // Add custom variables here:
 const variables = {
-  currentVersion: '2.0.1.4',
-  dotnetVersion: '6.0',
+  currentVersion: '1.1.13',
+  dotnetVersion: '5.0',
   zksnacksPublicKeyFingerprint: '6FB3 872B 5D42 292F 5992 0797 8563 4832 8949 861E'
 }
